LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts.html

<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<meta http-equiv="X-WebKit-CSP" content="sandbox allow-scripts">
This test passes if it does alert pass.
<script>
alert('PASS');
</script>
