third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-expected.txt

CONSOLE MESSAGE: line 8: PASS
This test passes if it does alert pass.
